add validação de administrador

+++ src/financas/controllers/lancamento.controller.ts
@@ -6,6 +6,7 @@ import {
   Param,
   Post,
   Put,
+  UnauthorizedException,
   UseGuards,
 } from '@nestjs/common';
 import { FinancasService } from '../services/financas.service';
@@ -14,6 +15,8 @@ import { Lancamento } from '../entities/lancamento.entity';
 import { TokenValidationGuard } from 'src/auth/guards/token-validation.guard';
 import { JwtAuthGuard } from 'src/auth/guards/jwt-auth.guard';
 import { CurrentUser } from 'src/auth/decorators/current-user.decorator';
+import { IsAdmin } from 'src/auth/decorators/is-admin.decorator';
+import { UpdateLancamentoDto } from '../dtos/UpdateLancamentoDto';
 
 @Controller('financas')
 export class LancamentosController {
@@ -41,7 +44,14 @@ export class LancamentosController {
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string, @CurrentUser('ongId') ongId: string) {
+  remove(
+    @Param('id') id: string,
+    @CurrentUser('ongId') ongId: string,
+    @IsAdmin('is_admin') isAdmin: boolean,
+  ) {
+    if (!isAdmin) {
+      throw new UnauthorizedException('Ação restrita para administradores');
+    }
     return this.financasService.remove(id, ongId);
   }
 
@@ -49,9 +59,10 @@ export class LancamentosController {
   @Put(':id')
   update(
     @Param('id') id: string,
-    @Body() lancamento: Lancamento,
+    @Body() lancamento: UpdateLancamentoDto,
     @CurrentUser('ongId') ongId: string,
   ) {
+    console.log(lancamento);
     return this.financasService.update(id, lancamento, ongId);
   }
 }
